No control workflow
Summary:
- Some redundant lines in snvCalling.sh is removed.
- For no control, now the control columns (INFO_control and ANNOTATION_control) are not annotated anymore.
- Confidence annotation script is completely ported in Python. Two previous separate scripts for with or without control are merged into one script.
- Now uses most recent dbSNP.
- Now use dbSNP-common instead, because dbSNP-All vcf file doesn't have AF information.
- INFO field of ExAC, EVS, and local control is always annotated to the VCF file (even with control).
- filter_PEoverlap.py now works without control column.
- Now filter_PEoverlap.py fetches data by using header name, instead of hardcoded index of header.

- Median is not calculated for no control. -t option should be set manually for confidence annotation.
- The options -m, -r, -e of new confidence annotation script are changed. The corresponding functions are turned on by default (without options), and by setting those options to turn off.

Test Plan: .

Reviewers: warsow, heinold, buchhalt

Reviewed By: buchhalt

Differential Revision: https://eilslabs-phabricator.dkfz.de/D214

+++ resources/analysisTools/snvPipeline/snvCalling.sh
@@ -72,9 +72,15 @@ then
 	exit 2
 fi
 
+if [[ ${GERMLINE_AVAILABLE} == "1" ]]; then
+    snvOut=${filenameMPileupOut}
+else
+    snvOut=${FILENAME_VCF_SNVS_TEMP}
+fi
+
 # filter out mutations with strand bias next to the motif GGnnGG (or CCnnCC if snv is on reverse strand)
 ${PERL_BINARY} ${TOOL_SEQ_CONTEXT_ANNOTATOR} ${FASTAFROMBED_BINARY} ${filenameMPileupTemp} ${REFERENCE_GENOME} 10 ${TOOLS_DIR} |
-${PYTHON_BINARY} ${TOOL_RAW_SNV_FILTER} --outf=${filenameMPileupOut} ${RAW_SNV_FILTER_OPTIONS} # --inf=$MPILEUP_SUBDIR/forStrandBiasFilter.${chr}.bed
+${PYTHON_BINARY} ${TOOL_RAW_SNV_FILTER} --outf=${snvOut} ${RAW_SNV_FILTER_OPTIONS} # --inf=$MPILEUP_SUBDIR/forStrandBiasFilter.${chr}.bed
 
 if [[ "$?" == 0 ]]
 then
@@ -85,50 +91,32 @@ else
 	exit 2
 fi
 
+if [[ ${GERMLINE_AVAILABLE} == "1" ]]; then
+    if [[ ${runCompareGermline} == true ]]; then
+        NP_MPILEUP=${SCRATCH_DIR}/NP_MPILEUP_CHR${chr}
+        mkfifo $NP_MPILEUP
 
-# after tumor mutations have been called, separation of SNVs and indels
-# if there is a germline BAM, first look up these positions in the control file by just piling up the bases, this is NO SNV calling
-errMessage=""
-removePileupOut=false
+        # if there is a germline BAM, first look up these positions in the control file by just piling up the bases, this is NO SNV calling
+        ${SAMTOOLS_BINARY} mpileup ${MPILEUPCONTROL_OPTS} -r ${chr} -l ${filenameMPileupOut} -f ${REFERENCE_GENOME} ${CONTROL_BAMFILE_FULLPATH_BP} > ${NP_MPILEUP} &
+        ${PERL_BINARY} ${TOOL_VCF_PILEUP_COMPARE} ${filenameMPileupOut} $NP_MPILEUP "Header" > ${FILENAME_VCF_SNVS_TEMP}
 
-NP_MPILEUP=${SCRATCH_DIR}/NP_MPILEUP_CHR${chr}
-NP_MPILEUP_SPLIT=${SCRATCH_DIR}/NP_MPILEUP_SPLIT_CHR${chr}
+        rm $NP_MPILEUP
 
-mkfifo $NP_MPILEUP $NP_MPILEUP_SPLIT
-
-if [[ ${runCompareGermline} == true ]]
-then
-    errMessage="vcf_pileup_compare pipe returned non-zero exit code; not moving tmp output files  (${FILENAME_VCF_SNVS_TEMP} to final name; moving $filenameMPileupOut to $MPILEUP_SUBDIR/"
-    removePileupOut=true
-
-    germlinebamfullpath=$CONTROL_BAMFILE_FULLPATH_BP
-
-
-    ${SAMTOOLS_BINARY} mpileup ${MPILEUPCONTROL_OPTS} -r ${chr} -l ${filenameMPileupOut} -f ${REFERENCE_GENOME} ${germlinebamfullpath} > ${NP_MPILEUP} &
-    ${PERL_BINARY} ${TOOL_VCF_PILEUP_COMPARE} ${filenameMPileupOut} $NP_MPILEUP "Header" | ${PERL_BINARY} ${TOOL_MPILEUP_SPLITTER} ${FILENAME_VCF_SNVS_TEMP} ${NP_MPILEUP_SPLIT} &
-
-else
-    errMessage="mpileupOutputSplitter returned non-zero exit code; not moving tmp output files (${FILENAME_VCF_SNVS_TEMP} to final name"
-
-    NP_MPILEUP_SPLIT_AMBIG=${SCRATCH_DIR}/NP_MPILEUP_SPLIT
-
-    cat $filenameMPileupOut | ${PERL_BINARY} ${TOOL_MPILEUP_SPLITTER} ${FILENAME_VCF_SNVS_TEMP} ${NP_MPILEUP_SPLIT} &
+        if [[ "$?" == 0  ]]; then
+            rm ${filenameMPileupOut}
+        else
+            echo "vcf_pileup_compare pipe returned non-zero exit code; not moving tmp output files  (${FILENAME_VCF_SNVS_TEMP} to final name; moving $filenameMPileupOut to $MPILEUP_SUBDIR/"
+            mv ${filenameMPileupOut} ${filenameMPileupError}
+            exit 2
+        fi
+    fi
 fi
 
-cat $NP_MPILEUP_SPLIT | ${PERL_BINARY} ${TOOL_AMBIGUOUS_STRIPPER} > /dev/null
+mv ${FILENAME_VCF_SNVS_TEMP} ${FILENAME_VCF_SNVS}
+touch ${FILENAME_VCF_SNVS_CHECKPOINT}
+[[ ${FILENAME_VCF_SNVS_CHECKPOINT_OLD} ]] && touch ${FILENAME_VCF_SNVS_CHECKPOINT_OLD}
 
-if [[ "$?" == 0  ]]
-then
-    mv ${FILENAME_VCF_SNVS_TEMP} ${FILENAME_VCF_SNVS}
-    touch ${FILENAME_VCF_SNVS_CHECKPOINT}
-    [[ ${FILENAME_VCF_SNVS_CHECKPOINT_OLD} ]] && touch ${FILENAME_VCF_SNVS_CHECKPOINT_OLD}
-    [[ ${runCompareGermline} == true ]] && rm ${filenameMPileupOut}
-    exit 0
-fi
+# Cleanup
+# [[ ${useCustomScratchDir} == true ]] && rm -rf ${SCRATCH_DIR}
 
-rm $NP_MPILEUP
-rm $NP_MPILEUP_SPLIT
-[[ ${useCustomScratchDir} == true ]] && rm -rf ${SCRATCH_DIR}
-echo "vcf_pileup_compare pipe returned non-zero exit code; not moving tmp output files  (${FILENAME_VCF_SNVS_TEMP} to final name; moving $filenameMPileupOut to $MPILEUP_SUBDIR/"
-[[ ${runCompareGermline} == true ]] && mv ${filenameMPileupOut} ${filenameMPileupError}
-exit 2
+exit 0
